backend/api/requirements.txt
absl-py==1.3.0
aiofiles==23.1.0
aiohttp==3.8.1
aiosignal==1.2.0
algoliasearch==2.6.3
algoliasearch-django==2.0.0
anyascii==0.3.2
anyio==3.6.2
appdirs==1.4.4
# apturl==0.5.2
asgiref==3.5.2
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
audioread==3.0.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
blinker==1.4
blis==0.7.9
breadability==0.1.20
# Brlapi==0.8.3
bs4==0.0.1
cachetools==5.2.0
catalogue==2.0.8
# ccsm==0.9.14.1
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
# chrome-gnome-shell==0.0.0
ci-info==0.3.0
# cliapp==1.20180812.1
click==8.0.3
# cmdtest==0.32+git
colorama==0.4.4
# command-not-found==0.3
# compizconfig-python==0.9.14.1
confection==0.0.4
configobj==5.0.8
configparser==5.3.0
coqpit==0.0.17
coverage==7.1.0
cryptography==40.0.2
# cupshelpers==1.0
cycler==0.11.0
cymem==2.0.7
dbus-python==1.2.18
decorator==4.4.2
# defer==1.0.6
defusedxml==0.7.1
discord==2.0.0
discord.py==2.0.1
distlib==0.3.6
# distro==1.7.0
# distro-info===1.1build1
Django==4.0.6
django-cors-headers==3.13.0
django-elasticsearch-dsl==7.3
django-js-asset==2.0.0
django-mptt==0.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docx==0.2.4
# duplicity==0.8.21
elastic-transport==8.4.0
elasticsearch==7.17.9
elasticsearch-dsl==7.4.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
etelemetry==0.3.0
exceptiongroup==1.1.0
fasteners==0.14.1
ffmpeg==1.4
filelock==3.9.0
fitz==0.0.1.dev2
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
flatbuffers==22.10.26
fonttools==4.37.1
frontend==0.0.3
frozenlist==1.3.1
fsspec==2023.4.0
future==0.18.2
gast==0.4.0
gensim==4.3.1
github==1.2.7
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gprof2dot==2022.7.29
greenlet==2.0.1
grpcio==1.50.0
gunicorn==20.1.0
# gyp==0.1?
h11==0.14.0
h5py==3.7.0
heroku==0.1.4
httplib2==0.20.2
huggingface-hub==0.13.4
idna==3.3
imageio==2.21.2
imageio-ffmpeg==0.4.7
importlib-metadata==4.6.4
iniconfig==2.0.0
ipaddress==1.0.23
iso8601==1.1.0
isodate==0.6.1
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==3.2.0
# kazam==1.4.5
keras==2.11.0
keyring==23.5.0
kiwisolver==1.4.4
langcodes==3.3.0
# language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.2
libclang==14.0.6
librosa==0.10.0.post2
llvmlite==0.39.1
lockfile==0.12.2
looseversion==1.1.2
# louis==3.20.0
lxml==4.9.2
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
memory-profiler==0.61.0
meson==1.0.1
monotonic==1.6
more-itertools==8.10.0
moviepy==1.0.3
msgpack==1.0.5
multidict==6.0.2
murmurhash==1.0.9
netifaces==0.11.0
networkx==3.1
nibabel==5.1.0
ninja==1.11.1
nipype==1.8.6
nltk==3.8.1
numba==0.56.4
numpy==1.23.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.0
olefile==0.46
oneai==0.9.53
opencv-python-headless==4.7.0.68
openshot-qt==3.0.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.4
paramiko==2.9.3
pathlib==1.0.1
pathy==0.10.1
pdfminer==20191125
pdfminer.six==20221105
pexpect==4.8.0
Pillow==9.0.1
pipenv==2022.9.21
platformdirs==2.6.1
pluggy==1.0.0
pooch==1.6.0
preshed==3.0.8
proglog==0.1.10
protobuf==3.12.4
prov==2.0.0
psutil==5.9.4
psycopg2==2.9.3
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycountry==22.3.5
pycparser==2.21
pycrc==0.10.0
pycryptodome==3.17
pycups==2.0.1
pydantic==1.10.7
pydot==1.4.2
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMuPDF==1.22.1
PyNaCl==1.5.0
pyparsing==2.4.7
PyPDF2==3.0.1
PyQt5==5.15.6
PyQt5-sip==12.9.1
PyQtWebEngine==5.15.5
pyRFC3339==1.1
pyrsistent==0.18.1
pyserial==3.5
pysftp==0.2.9
pytesseract==0.3.10
pytest==7.2.1
pytest-django==4.5.2
pytest-profiling==1.7.0
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-docx==0.8.11
python-dotenv==0.19.2
python-xlib==0.29
pythoncrc==1.21
pytz==2022.1
pyvesc==1.0.5
pyxdg==0.27
pyxnat==1.5
PyYAML==5.4.1
pyzmq==22.3.0
rdflib==6.3.2
regex==2023.3.23
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
scapy==2.5.0
scikit-learn==1.1.2
scipy==1.9.1
scp==0.14.4
SecretStorage==3.3.1
sentry-sdk==1.4.3
serial==0.0.97
simplejson==3.19.1
simplenote==2.1.4
six==1.16.0
sklearn==0.0
smart-open==6.3.0
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.3.2.post1
soxr==0.3.5
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==1.4.45
sqlparse==0.4.2
srsly==2.4.6
starlette==0.26.1
sumy==0.11.0
systemd-python==234
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
tensorflow-estimator==2.11.0
tensorflow-gpu==2.11.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.0
textblob==0.17.1
texttable==1.6.4
thinc==8.1.9
threadpoolctl==3.1.0
tokenizers==0.13.3
tomli==2.0.1
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.0
trainer==0.0.25
traits==6.3.2
transformers==4.28.1
ttystatus==0.38
typer==0.7.0
typing_extensions==4.3.0
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.21.1
validators==0.20.0
virtualenv==20.17.1
virtualenv-clone==0.5.7
wadllib==1.3.6
wasabi==1.1.1
websocket-client==1.2.3
websockets==10.4
Werkzeug==2.2.2
wrapt==1.14.1
xdg==5
xkit==0.0.0
yarl==1.8.1
zipp==1.0.0
